requirements.txt
asgiref==3.7.2
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudscraper==1.2.71
deckar01-ratelimit==3.0.2
deepl==1.15.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
eng-to-ipa==0.0.2
fonttools==4.42.1
futures==3.0.5
goslate==1.5.4
html5lib==1.1
idna==3.4
inflection==0.5.1
jellyfish==1.0.0
libretranslatepy==2.1.1
lxml==4.9.3
packaging==23.1
Pillow==10.0.0
PyDictionary==2.0.1
pyparsing==3.1.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
translate==3.6.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
webencodings==0.5.1
whitenoise==6.5.0
wordhoard==1.5.4
